Remove unused makefile target and fix wrong tcc test file size

+++ Makefile
@@ -1,7 +1,7 @@
 
 BENCH_LIBS := bench openlibm soft-fp
 
-$(BENCH_LIBS): %: latest
+$(BENCH_LIBS): %:
 	$(MAKE) -s -C ./src/common/$* archive
 
 COLOR_RED   = \033[1;31m
@@ -10,6 +10,10 @@ COLOR_NONE  = \033[0m
 
 RESULT = .result
 $(shell > $(RESULT))
+
+KEEP_LOG_FAILED ?= true
+KEEP_LOG_SUCCEED ?= false
+TIME := $(shell date --iso=seconds)
 
 ALL = mcf x264 tcc stream linpack gemm whetstone
 
@@ -22,17 +26,27 @@ all: $(BENCH_LIBS) $(ALL)
 		echo "====== Running OpenPerf [input *$${mainargs}*] ======"; \
 	fi
 
-$(ALL): %: $(BENCH_LIBS) latest
+$(ALL): %: $(BENCH_LIBS)
 	@{\
 		  TMP=$*.tmp;\
-	    make -C ./src/$* ARCH=$(ARCH) run 2>&1 | tee -a $$TMP;\
-      if [ $${PIPESTATUS[0]} -eq 0 ]; then \
+	    $(MAKE) -C ./src/$* ARCH=$(ARCH) run 2>&1 | tee -a $$TMP;\
+     if [ $${PIPESTATUS[0]} -eq 0 ]; then \
 		    printf "[%14s] $(COLOR_GREEN)PASS$(COLOR_NONE) " $* >> $(RESULT); \
 		    cat $$TMP | grep -E -i -e "OpenPerf time: ([0-9]*\.)?[0-9]*" >> $(RESULT); \
-		    rm $$TMP;\
+				if $(KEEP_LOG_SUCCEED); then \
+					mkdir -p "logs/$(TIME)/"; \
+					mv $$TMP "logs/$(TIME)/"; \
+				else \
+					rm $$TMP; \
+				fi \
 	    else \
 			  printf "[%14s] $(COLOR_RED)***FAIL***$(COLOR_NONE)\n" $* >> $(RESULT); \
-				rm $$TMP; \
+				if $(KEEP_LOG_FAILED); then \
+					mkdir -p "logs/$(TIME)/"; \
+					mv $$TMP "logs/$(TIME)/"; \
+				else \
+					rm $$TMP; \
+				fi \
 	    fi \
 	}
 
@@ -64,15 +78,11 @@ run: $(BENCH_LIBS) all
 	' $(RESULT)
 	@rm $(RESULT)
 
-libs: $(BENCH_LIBS)
-
 CLEAN_ALL = $(dir $(shell find . -mindepth 2 -name Makefile))
 clean-all: $(CLEAN_ALL)
 	 
 $(CLEAN_ALL):
 	-@$(MAKE) -s -C $@ clean
 
-.PHONY: $(BENCH_LIBS) $(CLEAN_ALL) $(ALL) all run clean-all latest libs
-
-latest:
+.PHONY: $(BENCH_LIBS) $(CLEAN_ALL) $(ALL) all run clean-all
